Updated notes controller impl with JWT token

+++ controllers/notes.js
@@ -3,7 +3,7 @@ const Note = require('../models/note');
 const response = require('../utils/response');
 
 exports.getNotes = (req, res) => {
-  Note.find({})
+  Note.find({ createdBy: req.user.id.toString() })
     .then((notes) => response.success(res, notes))
     .catch((err) => response.serverError(res, err));
 };
@@ -11,13 +11,14 @@ exports.getNotes = (req, res) => {
 exports.addNote = (req, res) => {
   const title = req.body.title;
   const description = req.body.description;
+  const createdBy = req.user.id;
 
   if (!title) {
     response.badRequest(res, 'Title is required to create note!');
     return;
   }
 
-  new Note({ title: title, description: description })
+  new Note({ title: title, description: description, createdBy: createdBy })
     .save()
     .then(() => response.success(res, 'Note added successfully'))
     .catch((err) => response.serverError(res, err, 'Error in adding note'));
@@ -28,9 +29,12 @@ exports.getNoteById = (req, res) => {
   if (!id || !ObjectId.isValid(id)) response.badRequest(res, 'Invalid note id!');
 
   Note.findById(id)
-    .then((success) => {
-      if (!success) response.notFound(res, `Note with id: ${id} was not found!`);
-      else response.success(res, success);
+    .then((note) => {
+      if (!note || note.createdBy.toString() != req.user.id.toString()) {
+        response.notFound(res, `Note with id: ${id} was not found!`);
+      } else {
+        response.success(res, note);
+      }
     })
     .catch((err) => {
       response.serverError(res, err, `Error in retrieving note with id: ${id}!`);
@@ -52,20 +56,26 @@ exports.updateNote = (req, res) => {
     return;
   }
 
-  Note.updateOne({ _id: id }, { title: title, description: description })
+  const updatedNote = { title: title };
+  if (description) updatedNote.description = description;
+
+  Note.updateOne({ _id: id, createdBy: req.user.id.toString() }, updatedNote)
     .then((success) => {
       console.log(success);
       if (!success) response.serverError(res, undefined, 'Something went wrong');
       if (success.modifiedCount == 0) response.notFound(res, `Can't find any note with ID: ${id}`);
       else response.success(res, `Successfully updated note with id: ${id}`);
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      console.log(err);
+      response.serverError(res, err, `Couldn't update!`);
+    });
 };
 
 exports.deleteNoteById = (req, res) => {
   const id = req.params.id;
   if (!id || !ObjectId.isValid(id)) return response.badRequest(res, `Invalid note ID: ${id}`);
-  Note.deleteOne({ _id: id })
+  Note.deleteOne({ _id: id, createdBy: req.user.id.toString() })
     .then((success) => {
       console.log(success);
       if (!success) response.serverError(res, undefined, 'Something went wrong');
